+++ api/gen/proto/go/teleport/devicetrust/v1/devicetrust_service_grpc.pb.go
@@ -18,7 +18,7 @@
 // - protoc             (unknown)
 // source: teleport/devicetrust/v1/devicetrust_service.proto
 
-package v1
+package devicetrustv1
 
 import (
 	context "context"