+++ tool/tsh/tsh_test.go
@@ -152,6 +152,9 @@ func (p *cliModules) AttestHardwareKey(_ context.Context, _ interface{}, _ keys.
 func (p *cliModules) EnableRecoveryCodes() {
 }
 
+func (p *cliModules) EnablePlugins() {
+}
+
 func TestAlias(t *testing.T) {
 	testExecutable, err := os.Executable()
 	require.NoError(t, err)
@@ -2771,7 +2774,7 @@ func TestSerializeDatabases(t *testing.T) {
 	    "redis": {}
 	  },
       "tls": {
-        "mode": "verify-full"
+        "mode": 0
       },
       "ad": {
         "domain": "",